@@ -1,11 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
-        minSdkVersion 7
+        minSdkVersion 14
         targetSdkVersion 23
         versionCode 1
         versionName "1.0"
@@ -19,6 +19,6 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:design:23.3.0'
+    compile 'com.google.android.material:material:1.0.0'
     compile project(':lib')
 }
