@@ -1,6 +1,6 @@
 package com.lsjwzh.adapter;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public interface OnRecyclerViewItemClickListener<VH extends RecyclerView.ViewHolder> {
