@@ -1,7 +1,7 @@
 package com.lsjwzh.widget.recyclerviewpagerdeomo;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public class SpacesItemDecoration extends RecyclerView.ItemDecoration {
