@@ -6,11 +6,14 @@ import java.util.Set;
 
 import android.annotation.TargetApi;
 import android.app.Activity;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.viewpager.widget.PagerAdapter;
+
 import android.util.Log;
 import android.util.SparseArray;
 import android.view.LayoutInflater;
@@ -18,8 +21,8 @@ import android.view.View;
 import android.view.ViewGroup;
 
 /**
- * Implementation of {@link android.support.v4.view.PagerAdapter} that
- * uses a {@link android.support.v4.app.Fragment} to manage each page. This class also handles
+ * Implementation of {@link PagerAdapter} that
+ * uses a {@link Fragment} to manage each page. This class also handles
  * saving and restoring of fragment's state.
  * <p/>
  * <p>This version of the pager is more useful when there are a large number
@@ -27,7 +30,7 @@ import android.view.ViewGroup;
  * the user, their entire fragment may be destroyed, only keeping the saved
  * state of that fragment.  This allows the pager to hold on to much less
  * memory associated with each visited page as compared to
- * {@link android.support.v4.app.FragmentPagerAdapter} at the cost of potentially more overhead when
+ * {@link FragmentPagerAdapter} at the cost of potentially more overhead when
  * switching between pages.
  * <p/>
  * <p>When using FragmentPagerAdapter the host ViewPager must have a
