@@ -1,7 +1,7 @@
 package com.lsjwzh.widget.recyclerviewpager;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.AttributeSet;
 import android.util.Log;
 
